+++ app/src/main/java/app/revanced/manager/ui/screen/InstalledAppInfoScreen.kt
@@ -32,6 +32,7 @@ import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.RadioButton
@@ -657,6 +658,18 @@ fun InstalledAppInfoScreen(
                 scrollBehavior = scrollBehavior,
                 onBackClick = {
                     if (viewModel.isInstalling) showLeaveInstallDialog = true else onBackClick()
+                },
+                actions = {
+                    if (viewModel.hasSavedCopy) {
+                        IconButton(
+                            onClick = { openExportPicker() }
+                        ) {
+                            Icon(
+                                imageVector = Icons.Outlined.Save,
+                                contentDescription = stringResource(R.string.export)
+                            )
+                        }
+                    }
                 }
             )
         },